+++ packages/cli/__tests__/__system__/define/urimap-client/cli.define.urimap.client.system.test.ts
@@ -16,12 +16,11 @@ let TEST_ENVIRONMENT: ITestEnvironment<ITestPropertiesSchema>;
 let certificate: string;
 
 describe("CICS define urimap-client command", () => {
-
   beforeAll(async () => {
     TEST_ENVIRONMENT = await TestEnvironment.setUp({
       testName: "define_urimap_client",
       installPlugin: true,
-      tempProfileTypes: ["cics"]
+      tempProfileTypes: ["cics"],
     });
     certificate = TEST_ENVIRONMENT.systemTestProperties.urimap.certificate;
   });
@@ -38,8 +37,16 @@ describe("CICS define urimap-client command", () => {
   });
 
   it("should get a syntax error if urimap name is omitted", () => {
-    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT,
-      ["", "FAKEGRP", "FAKEPATH", "FAKEHOST", "FAKERGN", "false", "BASIC", certificate]);
+    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT, [
+      "",
+      "FAKEGRP",
+      "FAKEPATH",
+      "FAKEHOST",
+      "FAKERGN",
+      "false",
+      "BASIC",
+      certificate,
+    ]);
     const stderr = output.stderr.toString();
     expect(stderr).toContain("Syntax");
     expect(stderr).toContain("urimap");
@@ -48,8 +55,16 @@ describe("CICS define urimap-client command", () => {
   });
 
   it("should get a syntax error if CSD group is omitted", () => {
-    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT,
-      ["FAKESRV", "", "FAKEPATH", "FAKEHOST", "FAKERGN", "false", "BASIC", certificate]);
+    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT, [
+      "FAKESRV",
+      "",
+      "FAKEPATH",
+      "FAKEHOST",
+      "FAKERGN",
+      "false",
+      "BASIC",
+      certificate,
+    ]);
     const stderr = output.stderr.toString();
     expect(stderr).toContain("Syntax");
     expect(stderr).toContain("csdGroup");
@@ -57,8 +72,16 @@ describe("CICS define urimap-client command", () => {
   });
 
   it("should get a syntax error if urimap path is omitted", () => {
-    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT,
-      ["FAKESRV", "FAKEGRP", "", "FAKEHOST", "FAKERGN", "false", "BASIC", certificate]);
+    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT, [
+      "FAKESRV",
+      "FAKEGRP",
+      "",
+      "FAKEHOST",
+      "FAKERGN",
+      "false",
+      "BASIC",
+      certificate,
+    ]);
     const stderr = output.stderr.toString();
     expect(stderr).toContain("Syntax");
     expect(stderr).toContain("urimap-path");
@@ -66,8 +89,16 @@ describe("CICS define urimap-client command", () => {
   });
 
   it("should get a syntax error if urimap host is omitted", () => {
-    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT,
-      ["FAKESRV", "FAKEGRP", "FAKEPATH", "", "FAKERGN", "false", "BASIC", certificate]);
+    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT, [
+      "FAKESRV",
+      "FAKEGRP",
+      "FAKEPATH",
+      "",
+      "FAKERGN",
+      "false",
+      "BASIC",
+      certificate,
+    ]);
     const stderr = output.stderr.toString();
     expect(stderr).toContain("Syntax");
     expect(stderr).toContain("urimap-host");
@@ -75,8 +106,16 @@ describe("CICS define urimap-client command", () => {
   });
 
   it("should get a syntax error if region name is omitted", () => {
-    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT,
-      ["FAKESRV", "FAKEGRP", "FAKEPATH", "FAKEHOST", "", "false", "BASIC", certificate]);
+    const output = runCliScript(__dirname + "/__scripts__/define_urimap_client.sh", TEST_ENVIRONMENT, [
+      "FAKESRV",
+      "FAKEGRP",
+      "FAKEPATH",
+      "FAKEHOST",
+      "",
+      "false",
+      "BASIC",
+      certificate,
+    ]);
     const stderr = output.stderr.toString();
     expect(stderr).toContain("Syntax");
     expect(stderr).toContain("region-name");