+++ packages/cli/__tests__/__unit__/define/urimap-pipeline/UrimapPipeline.handler.unit.test.ts
@@ -26,21 +26,21 @@ const protocol = "http";
 const rejectUnauthorized = false;
 
 const PROFILE_MAP = new Map<string, IProfile[]>();
-PROFILE_MAP.set(
-  "cics", [{
+PROFILE_MAP.set("cics", [
+  {
     name: "cics",
     type: "cics",
     host,
     port,
     user,
-    password
-  }]
-);
+    password,
+  },
+]);
 const PROFILES: CommandProfiles = new CommandProfiles(PROFILE_MAP);
 const DEFAULT_PARAMETERS: IHandlerParameters = mockHandlerParameters({
   positionals: ["cics", "define", "urimap-pipeline"],
   definition: UrimapPipelineDefinition,
-  profiles: PROFILES
+  profiles: PROFILES,
 });
 
 describe("DefineUrimapPipelineHandler", () => {
@@ -60,9 +60,9 @@ describe("DefineUrimapPipelineHandler", () => {
 
   const defaultReturn: ICMCIApiResponse = {
     response: {
-      resultsummary: {api_response1: "1024", api_response2: "0", recordcount: "0", displayed_recordcount: "0"},
-      records: "testing"
-    }
+      resultsummary: { api_response1: "1024", api_response2: "0", recordcount: "0", displayed_recordcount: "0" },
+      records: "testing",
+    },
   };
 
   const functionSpy = jest.spyOn(Define, "defineUrimapPipeline");
@@ -75,7 +75,7 @@ describe("DefineUrimapPipelineHandler", () => {
   it("should call the defineUrimapPipeline api with required options specified", async () => {
     const handler = new UrimapPipelineHandler();
 
-    const commandParameters = {...DEFAULT_PARAMETERS};
+    const commandParameters = { ...DEFAULT_PARAMETERS };
     commandParameters.arguments = {
       ...commandParameters.arguments,
       urimapName,
@@ -92,7 +92,7 @@ describe("DefineUrimapPipelineHandler", () => {
       user,
       password,
       rejectUnauthorized,
-      protocol
+      protocol,
     };
 
     await handler.process(commandParameters);
@@ -107,7 +107,7 @@ describe("DefineUrimapPipelineHandler", () => {
         user: testProfile.user,
         password: testProfile.password,
         rejectUnauthorized,
-        protocol
+        protocol,
       }),
       {
         name: urimapName,
@@ -121,7 +121,7 @@ describe("DefineUrimapPipelineHandler", () => {
         enable,
         description: undefined,
         transactionName: undefined,
-        webserviceName: undefined
+        webserviceName: undefined,
       }
     );
   });
@@ -129,7 +129,7 @@ describe("DefineUrimapPipelineHandler", () => {
   it("should call the defineUrimapPipeline api with all options specified", async () => {
     const handler = new UrimapPipelineHandler();
 
-    const commandParameters = {...DEFAULT_PARAMETERS};
+    const commandParameters = { ...DEFAULT_PARAMETERS };
     commandParameters.arguments = {
       ...commandParameters.arguments,
       urimapName,
@@ -149,7 +149,7 @@ describe("DefineUrimapPipelineHandler", () => {
       user,
       password,
       rejectUnauthorized,
-      protocol
+      protocol,
     };
 
     await handler.process(commandParameters);
@@ -164,7 +164,7 @@ describe("DefineUrimapPipelineHandler", () => {
         user: testProfile.user,
         password: testProfile.password,
         rejectUnauthorized,
-        protocol
+        protocol,
       }),
       {
         name: urimapName,
@@ -178,7 +178,7 @@ describe("DefineUrimapPipelineHandler", () => {
         transactionName,
         webserviceName,
         regionName,
-        cicsPlex
+        cicsPlex,
       }
     );
   });