+++ packages/cli/__tests__/__unit__/define/webservice/Webservice.handler.unit.test.ts
@@ -26,21 +26,21 @@ const protocol = "http";
 const rejectUnauthorized = false;
 
 const PROFILE_MAP = new Map<string, IProfile[]>();
-PROFILE_MAP.set(
-  "cics", [{
+PROFILE_MAP.set("cics", [
+  {
     name: "cics",
     type: "cics",
     host,
     port,
     user,
-    password
-  }]
-);
+    password,
+  },
+]);
 const PROFILES: CommandProfiles = new CommandProfiles(PROFILE_MAP);
 const DEFAULT_PARAMETERS: IHandlerParameters = mockHandlerParameters({
   positionals: ["cics", "define", "webservice"],
   definition: WebServiceDefinition,
-  profiles: PROFILES
+  profiles: PROFILES,
 });
 
 describe("DefineWebserviceHandler", () => {
@@ -53,9 +53,9 @@ describe("DefineWebserviceHandler", () => {
 
   const defaultReturn: ICMCIApiResponse = {
     response: {
-      resultsummary: {api_response1: "1024", api_response2: "0", recordcount: "0", displayed_recordcount: "0"},
-      records: "testing"
-    }
+      resultsummary: { api_response1: "1024", api_response2: "0", recordcount: "0", displayed_recordcount: "0" },
+      records: "testing",
+    },
   };
 
   const functionSpy = jest.spyOn(Define, "defineWebservice");
@@ -68,7 +68,7 @@ describe("DefineWebserviceHandler", () => {
   it("should call the defineWebservice api", async () => {
     const handler = new WebServiceHandler();
 
-    const commandParameters = {...DEFAULT_PARAMETERS};
+    const commandParameters = { ...DEFAULT_PARAMETERS };
     commandParameters.arguments = {
       ...commandParameters.arguments,
       webserviceName: websvcName,
@@ -82,7 +82,7 @@ describe("DefineWebserviceHandler", () => {
       user,
       password,
       rejectUnauthorized,
-      protocol
+      protocol,
     };
 
     await handler.process(commandParameters);
@@ -97,7 +97,7 @@ describe("DefineWebserviceHandler", () => {
         user: testProfile.user,
         password: testProfile.password,
         rejectUnauthorized,
-        protocol
+        protocol,
       }),
       {
         name: websvcName,
@@ -105,7 +105,7 @@ describe("DefineWebserviceHandler", () => {
         pipelineName,
         wsBind,
         regionName,
-        cicsPlex
+        cicsPlex,
       }
     );
   });