+++ packages/cli/__tests__/__unit__/enable/program/Program.handler.unit.test.ts
@@ -1,0 +1,102 @@
+/**
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ *
+ */
+
+import { mockHandlerParameters } from "@zowe/cli-test-utils";
+import { CommandProfiles, IHandlerParameters, IProfile, Session } from "@zowe/imperative";
+import { ICMCIApiResponse } from "../../../../src";
+import { ProgramDefinition } from "../../../../src/enable/program/Program.definition";
+import ProgramHandler from "../../../../src/enable/program/Program.handler";
+
+jest.mock("@zowe/cics-for-zowe-sdk");
+const Enable = require("@zowe/cics-for-zowe-sdk");
+
+const host = "somewhere.com";
+const port = "43443";
+const user = "someone";
+const password = "somesecret";
+const protocol = "http";
+const rejectUnauthorized = false;
+
+const PROFILE_MAP = new Map<string, IProfile[]>();
+PROFILE_MAP.set("cics", [
+  {
+    name: "cics",
+    type: "cics",
+    host,
+    port,
+    user,
+    password,
+    protocol,
+    rejectUnauthorized,
+  },
+]);
+const PROFILES: CommandProfiles = new CommandProfiles(PROFILE_MAP);
+const DEFAULT_PARAMETERS: IHandlerParameters = mockHandlerParameters({
+  positionals: ["cics", "enable", "program"],
+  definition: ProgramDefinition,
+  profiles: PROFILES,
+});
+
+describe("enableProgramHandler", () => {
+  const programName = "testProgram";
+  const regionName = "testRegion";
+
+  const defaultReturn: ICMCIApiResponse = {
+    response: {
+      resultsummary: { api_response1: "1024", api_response2: "0", recordcount: "0", displayed_recordcount: "0" },
+      records: "testing",
+    },
+  };
+
+  const functionSpy = jest.spyOn(Enable, "enableProgram");
+
+  beforeEach(() => {
+    functionSpy.mockClear();
+    functionSpy.mockImplementation(async () => defaultReturn);
+  });
+
+  it("should call the enableProgram api", async () => {
+    const handler = new ProgramHandler();
+
+    const commandParameters = { ...DEFAULT_PARAMETERS };
+    commandParameters.arguments = {
+      ...commandParameters.arguments,
+      name: programName,
+      regionName,
+      host,
+      port,
+      user,
+      password,
+      protocol,
+      rejectUnauthorized,
+    };
+
+    await handler.process(commandParameters);
+
+    expect(functionSpy).toHaveBeenCalledTimes(1);
+    const testProfile = PROFILE_MAP.get("cics")[0];
+    expect(functionSpy).toHaveBeenCalledWith(
+      new Session({
+        type: "basic",
+        hostname: testProfile.host,
+        port: testProfile.port,
+        user: testProfile.user,
+        password: testProfile.password,
+        rejectUnauthorized,
+        protocol,
+      }),
+      {
+        name: programName,
+        regionName,
+      }
+    );
+  });
+});