+++ packages/cli/src/define/urimap-client/UrimapClient.handler.ts
@@ -26,13 +26,12 @@ const strings = (require("../../-strings-/en").default as typeof i18nTypings).DE
  */
 export default class UrimapClientHandler extends CicsBaseHandler {
   public async processWithSession(params: IHandlerParameters, session: AbstractSession): Promise<ICMCIApiResponse> {
-
     const status: ITaskWithStatus = {
       statusMessage: "Defining URIMAP of type Client to CICS",
       percentComplete: 0,
-      stageName: TaskStage.IN_PROGRESS
+      stageName: TaskStage.IN_PROGRESS,
     };
-    params.response.progress.startBar({task: status});
+    params.response.progress.startBar({ task: status });
 
     const response = await defineUrimapClient(session, {
       name: params.arguments.urimapName,
@@ -45,7 +44,7 @@ export default class UrimapClientHandler extends CicsBaseHandler {
       regionName: params.arguments.regionName,
       cicsPlex: params.arguments.cicsPlex,
       authenticate: params.arguments.authenticate,
-      certificate: params.arguments.certificate
+      certificate: params.arguments.certificate,
     });
 
     params.response.console.log(strings.MESSAGES.SUCCESS, params.arguments.urimapName);