+++ packages/sdk/__tests__/__system__/define/Define.urimap-server.system.test.ts
@@ -29,11 +29,10 @@ function sleep(ms: number) {
 const sleepTime = 4000;
 
 describe("CICS Define server URImap", () => {
-
   beforeAll(async () => {
     testEnvironment = await TestEnvironment.setUp({
       testName: "cics_cmci_define_urimap-server",
-      tempProfileTypes: ["cics"]
+      tempProfileTypes: ["cics"],
     });
     csdGroup = testEnvironment.systemTestProperties.cmci.csdGroup;
     enable = false;
@@ -49,7 +48,7 @@ describe("CICS Define server URImap", () => {
       port: cicsProperties.port,
       type: "basic",
       rejectUnauthorized: cicsProperties.rejectUnauthorized || false,
-      protocol: cicsProperties.protocol as any || "https",
+      protocol: (cicsProperties.protocol as any) || "https",
     });
   });