+++ packages/sdk/src/doc/ICMCIApiResponse.ts
@@ -17,19 +17,19 @@ import { ICMCIResponseResultSummary } from "./ICMCIResponseResultSummary";
 
 export interface ICMCIApiResponse {
   /**
-     * See the following link for more information:
-     * https://www.ibm.com/support/knowledgecenter/SSGMCP_5.2.0/com.ibm.cics.ts.clientapi.doc/topics/clientapi_response_element.html
-     */
+   * See the following link for more information:
+   * https://www.ibm.com/support/knowledgecenter/SSGMCP_5.2.0/com.ibm.cics.ts.clientapi.doc/topics/clientapi_response_element.html
+   */
   response: {
     /**
-         * See the following link for more information:
-         * https://www.ibm.com/support/knowledgecenter/SSGMCP_5.2.0/com.ibm.cics.ts.clientapi.doc/topics/clientapi_resultsummary_element.html
-         */
+     * See the following link for more information:
+     * https://www.ibm.com/support/knowledgecenter/SSGMCP_5.2.0/com.ibm.cics.ts.clientapi.doc/topics/clientapi_resultsummary_element.html
+     */
     resultsummary: ICMCIResponseResultSummary;
     /**
-         * See the following link for more information:
-         * https://www.ibm.com/support/knowledgecenter/SSGMCP_5.2.0/com.ibm.cics.ts.clientapi.doc/topics/clientapi_records_element.html
-         */
+     * See the following link for more information:
+     * https://www.ibm.com/support/knowledgecenter/SSGMCP_5.2.0/com.ibm.cics.ts.clientapi.doc/topics/clientapi_records_element.html
+     */
     records: any;
   };
 }