+++ packages/sdk/src/methods/discard/Discard.ts
@@ -31,9 +31,17 @@ export async function discardProgram(session: AbstractSession, parms: IProgramPa
   Logger.getAppLogger().debug("Attempting to discard a program with the following parameters:\n%s", JSON.stringify(parms));
 
   const cicsPlex = parms.cicsPlex == null ? "" : parms.cicsPlex + "/";
-  const cmciResource = "/" + CicsCmciConstants.CICS_SYSTEM_MANAGEMENT + "/" +
-        CicsCmciConstants.CICS_PROGRAM_RESOURCE + "/" + cicsPlex + parms.regionName +
-        "?CRITERIA=(PROGRAM=" + parms.name + ")";
+  const cmciResource =
+    "/" +
+    CicsCmciConstants.CICS_SYSTEM_MANAGEMENT +
+    "/" +
+    CicsCmciConstants.CICS_PROGRAM_RESOURCE +
+    "/" +
+    cicsPlex +
+    parms.regionName +
+    "?CRITERIA=(PROGRAM=" +
+    parms.name +
+    ")";
   return CicsCmciRestClient.deleteExpectParsedXml(session, cmciResource, []);
 }
 
@@ -54,9 +62,17 @@ export async function discardTransaction(session: AbstractSession, parms: ITrans
   Logger.getAppLogger().debug("Attempting to discard a transaction with the following parameters:\n%s", JSON.stringify(parms));
 
   const cicsPlex = parms.cicsPlex == null ? "" : parms.cicsPlex + "/";
-  const cmciResource = "/" + CicsCmciConstants.CICS_SYSTEM_MANAGEMENT + "/" +
-        CicsCmciConstants.CICS_LOCAL_TRANSACTION + "/" + cicsPlex + parms.regionName +
-        "?CRITERIA=(TRANID=" + parms.name + ")";
+  const cmciResource =
+    "/" +
+    CicsCmciConstants.CICS_SYSTEM_MANAGEMENT +
+    "/" +
+    CicsCmciConstants.CICS_LOCAL_TRANSACTION +
+    "/" +
+    cicsPlex +
+    parms.regionName +
+    "?CRITERIA=(TRANID=" +
+    parms.name +
+    ")";
   return CicsCmciRestClient.deleteExpectParsedXml(session, cmciResource, []);
 }
 
@@ -67,8 +83,13 @@ export async function discardUrimap(session: AbstractSession, parms: IURIMapParm
   Logger.getAppLogger().debug("Attempting to discard a URIMap with the following parameters:\n%s", JSON.stringify(parms));
 
   const cicsPlex = parms.cicsPlex == null ? "" : parms.cicsPlex + "/";
-  const cmciResource = "/" + CicsCmciConstants.CICS_SYSTEM_MANAGEMENT + "/" +
-        CicsCmciConstants.CICS_URIMAP + "/" + cicsPlex +
-        `${parms.regionName}?CRITERIA=(NAME='${parms.name}')`;
+  const cmciResource =
+    "/" +
+    CicsCmciConstants.CICS_SYSTEM_MANAGEMENT +
+    "/" +
+    CicsCmciConstants.CICS_URIMAP +
+    "/" +
+    cicsPlex +
+    `${parms.regionName}?CRITERIA=(NAME='${parms.name}')`;
   return CicsCmciRestClient.deleteExpectParsedXml(session, cmciResource, []);
 }