+++ packages/vsce/src/trees/CICSPlexTree.ts
@@ -57,8 +57,9 @@ export class CICSPlexTree extends TreeItem {
     this.resourceFilters = {};
     this.activeFilter = undefined;
     this.groupName = group;
-    this.iconPath = group
-      ? getIconPathInResources("cics-system-group-dark.svg ", "cics-system-group-light.svg ")
+    this.iconPath =
+      group ?
+        getIconPathInResources("cics-system-group-dark.svg ", "cics-system-group-light.svg ")
       : getIconPathInResources("cics-plex-dark.svg", "cics-plex-light.svg");
   }