+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the IBM® CICS® Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+
+- Enhacement: Added the following commands: [#14](https://github.com/zowe/cics-for-zowe-client/issues/14)
+  - `zowe cics enable transaction`
+  - `zowe cics enable program`
+  - `zowe cics enable local-file`
+
 ## `6.0.0-next.202403042201`
 
 - V3 Breaking: Increased Node Engine to 18.12.0